Move the alt-svc logic to the H2 connection factory
