+++ jetty-core/jetty-http3/jetty-http3-server/src/main/java/org/eclipse/jetty/http3/server/AbstractHTTP3ServerConnectionFactory.java
@@ -83,11 +83,6 @@ public abstract class AbstractHTTP3ServerConnectionFactory extends AbstractConne
         this.connector = connector;
     }
 
-    protected Connector getConnector()
-    {
-        return connector;
-    }
-
     @Override
     protected void doStart() throws Exception
     {
