+++ jetty-core/jetty-http3/jetty-http3-server/src/main/java/org/eclipse/jetty/http3/server/HTTP3ServerConnectionFactory.java
@@ -17,9 +17,6 @@ import java.util.Objects;
 import java.util.concurrent.Executor;
 import java.util.concurrent.TimeoutException;
 
-import org.eclipse.jetty.http.HttpFields;
-import org.eclipse.jetty.http.HttpHeader;
-import org.eclipse.jetty.http.HttpVersion;
 import org.eclipse.jetty.http3.HTTP3Stream;
 import org.eclipse.jetty.http3.api.Session;
 import org.eclipse.jetty.http3.api.Stream;
@@ -28,11 +25,7 @@ import org.eclipse.jetty.http3.server.internal.HTTP3SessionServer;
 import org.eclipse.jetty.http3.server.internal.HTTP3StreamServer;
 import org.eclipse.jetty.http3.server.internal.HttpStreamOverHTTP3;
 import org.eclipse.jetty.http3.server.internal.ServerHTTP3StreamConnection;
-import org.eclipse.jetty.server.ConnectionMetaData;
-import org.eclipse.jetty.server.Connector;
 import org.eclipse.jetty.server.HttpConfiguration;
-import org.eclipse.jetty.server.NetworkConnector;
-import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.util.Promise;
 import org.eclipse.jetty.util.thread.Invocable;
 import org.eclipse.jetty.util.thread.ThreadPool;
@@ -49,27 +42,6 @@ public class HTTP3ServerConnectionFactory extends AbstractHTTP3ServerConnectionF
     public HTTP3ServerConnectionFactory(HttpConfiguration configuration)
     {
         super(configuration, new HTTP3SessionListener());
-        configuration.addCustomizer(new AltSvcCustomizer());
-    }
-
-    private class AltSvcCustomizer implements HttpConfiguration.Customizer
-    {
-        @Override
-        public Request customize(Request request, HttpFields.Mutable responseHeaders)
-        {
-            ConnectionMetaData connectionMetaData = request.getConnectionMetaData();
-            if (HttpVersion.HTTP_2 == connectionMetaData.getHttpVersion())
-            {
-                Connector h3Connector = HTTP3ServerConnectionFactory.this.getConnector();
-                if (h3Connector instanceof NetworkConnector nc)
-                {
-                    int port = nc.getLocalPort();
-                    if (port > 0)
-                        responseHeaders.add(HttpHeader.ALT_SVC, String.format("h3=\":%d\"", port));
-                }
-            }
-            return request;
-        }
     }
 
     private static class HTTP3SessionListener implements HTTP3SessionServer.Listener
